dr_blog/requirements.txt
appdirs==1.4.4
asgiref==3.3.4
certifi==2020.6.20
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
cryptography==3.4.7
defusedxml==0.7.1
Deprecated==1.2.12
distlib==0.3.1
Django==3.2.3
django-cors-headers==3.7.0
django-oauth-toolkit==1.5.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.4.0
drf-social-oauth2==1.0.9
filelock==3.0.12
idna==2.10
importlib-metadata==1.7.0
importlib-resources==3.0.0
itypes==1.2.0
Jinja2==3.0.1
jwcrypto==0.9.1
MarkupSafe==2.0.1
oauthlib==3.1.1
pbr==5.4.5
Pillow==8.2.0
pipenv==2020.6.2
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
selenium==3.141.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
stevedore==2.0.0
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.25.10
virtualenv==20.0.30
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
wrapt==1.12.1
zipp==3.1.0
